pyproject.toml
[tool.poetry]
name = "smartroute"
version = "0.1.0"
description = "This is a RouterAI module, which integrates multiple AI APIs and routes requests to provide users with the best service for their need"
authors = ["Daniel Rezende <danielrezende2@academico.ufs.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "0.115.11" }
langchain = "^0.3.20"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.8.1"
langchain-openai = "^0.3.8"
langchain-google-genai = "^2.0.11"
langchain-anthropic = "^0.3.9"
torch = "^2.6.0"
transformers = "^4.49.0"
psycopg = {extras = ["binary"], version = "^3.2.5"}
langchain-postgres = "^0.0.13"
pyjwt = "^2.10.1"
alembic = "^1.15.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.9.10"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
